Move Diggy experience to common features/gui
